Extract WTA tab state and input editing
Move per-tab state and pure input editing behavior out of app.rs while preserving existing app re-exports and runtime behavior.

Copilot-Session: 725545ce-7285-44f9-ab47-538b3149d444
